refactor: bring back stac-validate
