+++ crates/core/Cargo.toml
@@ -55,6 +55,7 @@ url = { workspace = true, features = ["serde"] }
 assert-json-diff.workspace = true
 bytes.workspace = true
 rstest.workspace = true
+stac-validate.workspace = true
 tokio = { workspace = true, features = ["macros"] }
 tokio-test.workspace = true
