+++ crates/core/src/lib.rs
@@ -290,6 +290,7 @@ pub fn version() -> &'static str {
 #[cfg(test)]
 mod tests {
     use rstest as _;
+    use stac_validate as _;
     use tokio as _;
     use tokio_test as _;
