+++ crates/duckdb/Cargo.toml
@@ -33,4 +33,4 @@ thiserror.workspace = true
 [dev-dependencies]
 geo.workspace = true
 rstest.workspace = true
-stac-io = { workspace = true, features = ["validate"] }
+stac-validate.workspace = true
