+++ crates/validate/src/lib.rs
@@ -6,7 +6,7 @@
 //!
 //! ```
 //! use stac::Item;
-//! use stac_io::Validate;
+//! use stac_validate::Validate;
 //!
 //! Item::new("an-id").validate().unwrap();
 //! ```
@@ -16,7 +16,7 @@
 //!
 //! ```
 //! # use stac::Item;
-//! use stac_io::Validator;
+//! use stac_validate::Validator;
 //! let mut items: Vec<_> = (0..10).map(|n| Item::new(format!("item-{}", n))).collect();
 //! let mut validator = Validator::new().unwrap();
 //! for item in items {
@@ -27,20 +27,22 @@
 //! [Validator] is cheap to clone, so you are encouraged to validate a large
 //! number of objects at the same time if that's your use-case.
 
-use crate::Result;
 use serde::Serialize;
 
+mod error;
 mod validator;
 
-pub use validator::Validator;
+pub use {error::Error, validator::Validator};
+
+/// Public result type.
+pub type Result<T> = std::result::Result<T, Error>;
 
 /// Validate any serializable object with [json-schema](https://json-schema.org/)
 pub trait Validate: Serialize + Sized {
     /// Validates this object.
     ///
-    /// If the object fails validation, this will return an
-    /// [Error::Validation](crate::Error::Validation) which contains a vector of
-    /// all of the validation errors.
+    /// If the object fails validation, this will return an [Error::Validation]
+    /// which contains a vector of all of the validation errors.
     ///
     /// If you're doing multiple validations, use [Validator::validate], which
     /// will re-use cached schemas.
@@ -49,7 +51,7 @@ pub trait Validate: Serialize + Sized {
     ///
     /// ```
     /// use stac::Item;
-    /// use stac_io::Validate;
+    /// use stac_validate::Validate;
     ///
     /// let mut item = Item::new("an-id");
     /// item.validate().unwrap();
@@ -61,3 +63,8 @@ pub trait Validate: Serialize + Sized {
 }
 
 impl<T: Serialize> Validate for T {}
+
+/// Returns a string suitable for use as a HTTP user agent.
+pub fn user_agent() -> &'static str {
+    concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"))
+}
