+++ crates/validate/src/schemas/geojson/Feature.json
