+++ crates/validate/src/schemas/geojson/Geometry.json
