+++ crates/validate/src/schemas/json-schema/draft-07.json
