+++ crates/validate/src/schemas/v1.0.0/basics.json
