+++ crates/validate/src/schemas/v1.1.0/basics.json
